requirements.txt
python-dotenv==1.0.0
httpx==0.26.0
mcp-server==0.1.0 